Clean up app.js and created validator.js
